docs/newarchitecture/gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
prometheus-client==0.20.0
PyJWT[crypto]==2.9.0
cryptography==43.0.0
pydantic==2.7.4
pydantic-settings==2.3.4
PyYAML==6.0.1
SQLAlchemy==2.0.31
asyncpg==0.29.0
orjson==3.10.6
